ODFs and Tractography (#297)
* Init DTI/DKI fODFs computation file

* Format for readibility

* Add spherical sampling function

* Init ODF module

* Convert DSI MATLAB directions to Python

* Remove redundant import from sphericalsampling

* Add functions to convert ODF to spherical or cartesian form

* Expand function documentation

* Update function sphericalsampling to include ODF faces

* Add QOL revisions to sphericalsampling

* Complete main ODF file

* Change DSI grid sampling from MATLAB to Python

* Add DTI ODF calculation

* Add check for nan and inf spherical harmonics

* Remove nan check from DSI

* Fix error in computation of cartesian kODF

* Add ODF scaling

* Fix zero ODF scaling when no map is provided

* Fix incorrect computation of B12 coefficient

* Add tensor regularization for high FA values

* Add function to sample max ODF coefficient on sphere

* Invert (-1)^m term in SH coefficient

* Add minimum zero threshold

* Add documentation for tensor regularization

* Remove scaling normalization

* Fix inputs to DTI ODF functions

* Remove debugging functions and update doc

* Remove debug functions

* Add SVD inverse fallback

* Incorporate ODF to main

* Fix for incorrect FBI ODF directions

* Invert z-axis for DSI Studio

* Fix direction invariance between FBI and DTI/DKI ODF

* Update fib file creation

* Add correct method for importing QA map

* Move tractography from main() to fit_regime()

* Add ODF scaling for DSI Studio visualization

* Update SH expansion to correctly account for Condon–Shortley phase

* Add Tournier and Descoteaux SH basis sets for MRtrix compatibility

* Revert DKI FT map to FA

* Add Fortran-like indexing to `utils.vectorize()`

* Fix faces return shape in sphericalsampling

* Disable reorientation of DWI to RAS

Fixes the issue of FSL gradients from not being rotates. It's better to leave the DWI and gradients in their original form.

* Fix Condon–Shortley phase in SH basis creation

* Add data save type odf.savenii()

* Add option to import other maps into fib file

* Update other_map import logic in dsistudio.py

* Add other maps to fit_regime

* Replace FBI shbasis with tournier

* Update fnames

* Code cleanup

* Add try-error in ODF computation to catch invalid values

* Add proper __init__ statements to make modules installable

* Update des2dke.m for compatibility

* Update version and software info

* Update CHANGELOG

* Update documentation

+++ designer/tractography/dsistudio.py
@@ -18,8 +18,9 @@ import numpy as np
 import nibabel as nib
 from dipy.core.geometry import cart2sphere
 from dipy.core.sphere import HemiSphere
-from dipy.direction import peak_directions
+from dipy.direction import peak_directions, gfa
 from scipy.io.matlab import loadmat, savemat
+from designer.tractography import sphericalsampling
 from tqdm import tqdm
 ODF_COLS = 20000  # Number of columns in DSI Studio odf split
 tqdmWidth = 70
@@ -79,7 +80,7 @@ def convertLPS(input, output):
         '-quiet',
         '-force',
         input,
-        '-strides', '-1,-2,3',
+        '-strides', '-1,-2,3,4',
         output
     ]
     completion = subprocess.run(arg)
@@ -87,7 +88,8 @@ def convertLPS(input, output):
         raise Exception('Conversion of NifTI file to LPS failed. '
         'Check above for errors.')
 
-def makefib(input, output, mask=None, n_fibers=3):
+def makefib(input, output, map=None, mask=None, n_fibers=5, scale=1,
+    other_maps=None):
     """
     Converts a NifTi ``.nii`` file containing sh coefficients to a DSI
     Studio fib file
@@ -102,11 +104,25 @@ def makefib(input, output, mask=None, n_fibers=3):
         Path to input nifti file containing SH coefficients
     output : str
         Path to output fib file; must end with .fib
+    map : str
+        Path to dMRI map to control stopping criteria; usually an FA map. This
+        map is stored under variable fa_n (n =  1, 2,..., n ) in .fib file.
+        This maps is visualized by qa in DSI Studio. The peak amplitude response
+        is used if no map is provided
     mask : str, optional
         Path to nifti file containing brain mask
     n_fibers : int, optional
         The maximum number ODF maxima to extract per voxel
         (Default: 3)
+    scale: float; optional 
+        Affects overall size of odfs in .fib file. This only affects
+        visualization.
+        (Default: 1)
+    other_maps : list of str; optional
+        List containing paths to other metrics maps to include into the .fib
+        file for computing metrics values along tracks. The maps could be AD,
+        AK, MK, AWF, etc. as long as they are the same size as SH coefficient
+        file.
 
     Returns
     -------
@@ -124,29 +140,46 @@ def makefib(input, output, mask=None, n_fibers=3):
         '(.nii)')
     if op.splitext(output)[-1] != '.fib':
         raise IOError('Output file needs to specified as a .fib file')
+    if not map is None:
+        if not op.exists(map):
+            raise OSError('Path to map image does not exist. Please '
+            'ensure that the folder specified exists.')
     if not mask is None:
         if not op.exists(mask):
             raise OSError('Path to brain mask does not exist. Please '
             'ensure that the folder specified exists.')
+    if type(other_maps) == list or other_maps == None:
+        if type(other_maps) == list:
+            if any([not op.exists(x) for x in other_maps]):
+                raise OSError('One of the paths defined in other maps does not '
+                'exist Please ensure all files exist.')
+    else:
+        raise TypeError('Path to other maps needs to be entered as a list of '
+        'strings defining paths to other metric files.')
     outdir = op.dirname(output)
     # Convert to LPS
     fname, ext = op.splitext(op.basename(input))
     fname = fname +'_lps'
     input_ = op.join(outdir, fname + ext)
+    convertLPS(input, input_)
+    if not map is None:
+        fname, ext = op.splitext(op.basename(map))
+        fname = fname +'_lps'
+        map_ = op.join(outdir, fname + ext)
+        convertLPS(map, map_)
     if not mask is None:
-        (fname, ext) = op.splitext(op.basename(mask))
+        fname, ext = op.splitext(op.basename(mask))
         fname = fname +'_lps'
         mask_ = op.join(outdir, fname + ext)
-    convertLPS(input, input_)
-    convertLPS(mask, mask_)
+        convertLPS(mask, mask_)
     # Get ODF geometry
-    verts, faces = get_dsi_studio_ODF_geometry("odf8")
+    verts, faces = sphericalsampling.dsigrid('odf8')
     num_dirs, _ = verts.shape
     hemisphere = num_dirs // 2
     x, y, z = verts[:hemisphere].T
     hs = HemiSphere(x=x, y=y, z=z)
     # Convert to DSI Studio LPS+ from MRTRIX3 RAS+
-    _, theta, phi = cart2sphere(x, y, z)
+    _, theta, phi = cart2sphere(-x, -y, z)
     dirs_txt = op.join(outdir, "directions.txt")
     np.savetxt(dirs_txt, np.column_stack([phi, theta]))
     # Get SH amplitude
@@ -179,10 +212,48 @@ def makefib(input, output, mask=None, n_fibers=3):
         ampl_data.shape[2]), order='F')
     # Make flat mask
     flat_mask = mask_img.flatten(order='F') > 0
-    odf_array = ampl_data.reshape(-1, ampl_data.shape[3], order="F")
+    odf_array = ampl_data.reshape(-1, ampl_data.shape[3], order='F')
     masked_odfs = odf_array[flat_mask, :]
+    masked_odfs = np.nan_to_num(masked_odfs)
+    if not map is None:
+        map_img = nib.load(map_)
+        if not np.allclose(map_img.affine, amplitudes_img.affine):
+            raise ValueError('Differing orientation between map image and '
+            'amplitudes.')
+        if not map_img.shape == amplitudes_img.shape[:3]:
+            raise ValueError('Differing grid between map image and '
+            'amplitudes')
+        map_img = map_img.get_fdata().flatten(order='F')
+        map_img[map_img < 0] = 0
+        masked_map = map_img[flat_mask]
+
+    # Scale ODF
+    masked_odfs = scale * masked_odfs
+
+    # Compute GFA
+    masked_gfa = np.zeros(masked_odfs.shape[0])
+    for vox in tqdm(range(masked_odfs.shape[0]),
+        desc='Computing GFA',
+        bar_format='{desc}: [{percentage:0.0f}%]',
+        unit='vox',
+        ncols=tqdmWidth):
+        masked_gfa[vox] = gfa(masked_odfs[vox, :])
+    
+    # Remove voxels possibly containing errors in ODF calculations. This is
+    # particularly important for volumes that were co-registered as it fixes
+    # large amplitudes ODF amplitude at border voxels
+    if not map is None:
+        idx = np.logical_or(masked_map == 0, np.sum(masked_odfs, axis=1) == 0)
+        masked_map[idx] = 0
+    else:
+        idx = np.sum(masked_odfs, axis=1) == 0
+    masked_odfs[idx, :] = 0
+    masked_gfa[idx] = 0
+    if not map is None:
+        masked_map[idx] = 0 
+
     n_odfs = masked_odfs.shape[0]
-    peak_indices = np.zeros((n_odfs, n_fibers))
+    peak_indices = np.zeros((n_odfs, n_fibers), dtype=int)
     peak_vals = np.zeros((n_odfs, n_fibers))
     dsi_mat = {}
     # Create matfile that can be read by dsi Studio
@@ -203,7 +274,12 @@ def makefib(input, output, mask=None, n_fibers=3):
     for nfib in range(n_fibers):
         # fill in the "fa" values
         fa_n = np.zeros(n_voxels)
-        fa_n[flat_mask] = peak_vals[:, nfib]
+        if not map is None:
+            fa_fromidx = np.zeros(masked_map.shape)
+            np.putmask(fa_fromidx, peak_vals[:, nfib] != 0, masked_map)
+            fa_n[flat_mask] = fa_fromidx
+        else:
+            fa_n[flat_mask] = peak_vals[:, nfib]
         dsi_mat['fa%d' % nfib] = fa_n.astype(np.float32)
         # Fill in the index values
         index_n = np.zeros(n_voxels)
@@ -218,9 +294,30 @@ def makefib(input, output, mask=None, n_fibers=3):
     dsi_mat['odf_vertices'] = verts.T
     dsi_mat['odf_faces'] = faces.T
     dsi_mat['z0'] = np.array([1.])
+    # Fill in GFA
+    g_fa = np.zeros(n_voxels)
+    g_fa[flat_mask] = masked_gfa
+    dsi_mat['gfa'] = g_fa.astype(np.float32)
+    # Fill in other metrics maps
+    if not other_maps is None:
+        for path_map in other_maps:
+            map_name, ext = op.splitext(op.basename(path_map))
+            map_dir = op.dirname(path_map)
+            map_lps = op.join(map_dir, map_name + '_lps' + ext)
+            convertLPS(path_map, map_lps)
+            other_img = nib.load(map_lps)
+            if not other_img.shape == amplitudes_img.shape[:3]:
+                raise ValueError('Differing grid between other map image: {} '
+                'and amplitudes'.format(path_map))
+            gmap = other_img.get_fdata().flatten(order='F')
+            dsi_mat[map_name] = gmap.astype(np.float32)
+            os.remove(map_lps)
     savemat(output, dsi_mat, format='4', appendmat=False)
     # Remove unwanted files
     os.remove(input_)
-    os.remove(mask_)
+    if not mask is None:
+        os.remove(mask_)
+    if not map is None:
+        os.remove(map_)
     os.remove(dirs_txt)
     os.remove(odf_amplitudes_nii)
